+++ Example/DrawView-Example/ViewController.m
@@ -8,10 +8,15 @@
 
 #import "ViewController.h"
 #import <DrawView.h>
-@interface ViewController () {
+#import <AssetsLibrary/AssetsLibrary.h>
+
+@interface ViewController () <UIActionSheetDelegate> {
     IBOutlet DrawView *drawingView;
 }
 - (IBAction)loadArchived:(id)sender;
+- (IBAction)saveDrawing:(id)sender;
+- (IBAction)animateDrawing:(id)sender;
+- (IBAction)signatureMode:(id)sender;
 @end
 
 @implementation ViewController
@@ -20,13 +25,13 @@
     [super viewDidLoad];
 	// Do any additional setup after loading the view, typically from a nib.
     self.navigationItem.title = @"Drawing View";
-    UIBarButtonItem *animateButton = [[UIBarButtonItem alloc] initWithTitle:@"Animate" style:UIBarButtonItemStylePlain target:drawingView action:@selector(animatePath)];
+    UIBarButtonItem *animateButton = [[UIBarButtonItem alloc] initWithTitle:@"Clear" style:UIBarButtonItemStylePlain target:drawingView action:@selector(clearDrawing)];
     self.navigationItem.rightBarButtonItem = animateButton;
     UIBarButtonItem *archivedButton = [[UIBarButtonItem alloc] initWithTitle:@"Load" style:UIBarButtonItemStylePlain target:self action:@selector(loadArchived:)];
     self.navigationItem.leftBarButtonItem = archivedButton;
     // Drawing view setup.
-    [drawingView setBackgroundColor:[UIColor whiteColor]];
-    [drawingView strokeColor:[UIColor blackColor]];
+    drawingView.strokeColor = [UIColor redColor];
+    drawingView.strokeWidth = 25.0f;
 }
 - (void)didReceiveMemoryWarning{
     [super didReceiveMemoryWarning];
@@ -34,14 +39,40 @@
 }
 - (IBAction)loadArchived:(id)sender{
     // Load an archived array of bezier paths
-    UIBezierPath *bezPath = [UIBezierPath new];
-    NSData *testPath = [NSData dataWithContentsOfFile:[[NSBundle mainBundle] pathForResource:@"test" ofType:@"txt"]];
-    NSArray *paths = [NSKeyedUnarchiver unarchiveObjectWithData:testPath];
-    for (UIBezierPath *path in paths){
-        [bezPath appendPath:path];
-    }
+    UIBezierPath *path = [NSKeyedUnarchiver unarchiveObjectWithFile:[[NSBundle mainBundle] pathForResource:@"test-path" ofType:@"txt"]];
     // Display archived path.
-    [drawingView setDebugBox:YES];
-    [drawingView drawBezier:bezPath];
+    [drawingView drawBezier:path];
+}
+- (IBAction)animateDrawing:(id)sender{
+    [drawingView animatePath];
+}
+- (IBAction)saveDrawing:(id)sender{
+    UIActionSheet *saveSheet = [[UIActionSheet alloc] initWithTitle:@"Save" delegate:self cancelButtonTitle:@"Cancel" destructiveButtonTitle:nil otherButtonTitles:@"Camera Roll",@"UIImage",@"UIBezierPath", nil];
+    [saveSheet showInView:self.view];
+}
+- (IBAction)signatureMode:(id)sender{
+    [drawingView setMode:SignatureMode];
+}
+- (void)didRotateFromInterfaceOrientation:(UIInterfaceOrientation)fromInterfaceOrientation{
+    [drawingView refreshCurrentMode];
+}
+#pragma mark - UIActionSheet Delegate
+- (void)actionSheet:(UIActionSheet *)actionSheet didDismissWithButtonIndex:(NSInteger)buttonIndex{
+    if (buttonIndex != actionSheet.cancelButtonIndex){
+        if (buttonIndex == 0){
+            UIImage *drawingImage = [drawingView imageRepresentation];
+            ALAssetsLibrary *library = [[ALAssetsLibrary alloc] init];
+            [library writeImageToSavedPhotosAlbum:drawingImage.CGImage orientation:ALAssetOrientationUp completionBlock:^(NSURL *assetURL, NSError *error) {
+                NSLog(@"%@",assetURL);
+                NSLog(@"%@",error);
+            }];
+        }else if (buttonIndex == 1){
+            UIImage *drawingImage = [drawingView imageRepresentation];
+            NSLog(@"%@",drawingImage);
+        }else if (buttonIndex == 2){
+            UIBezierPath *path = [drawingView bezierPathRepresentation];
+            NSLog(@"%@",path);
+        }
+    }
 }
 @end