+++ CHANGELOG.md
@@ -3,3 +3,14 @@
 ## 0.1.0
 
 Initial release.
+
+## 0.2.0
+
+CocoaPod release
+
+## 1.0.0
+
+- Support to read out the current path in the drawing view.
+- Updated example showing how to read the current drawing view.
+- Added signature mode.
+- Ability to clear drawing view.